move share button to ContentContainerViewModel and also add icon for change language
